Filtration by Log2FoldChange
This was the first step in distinguishing which G-quadruplexes are to be focused on, on the basis of the log2FoldChange values. It is important to understand that the dataset could be full of false positives. Thus, one would have to check manually to check for the same or could simply increase the stringency of the filtration.